+++ packages/asc-ui/src/components/Toggle/index.ts
@@ -1,5 +1,1 @@
-export {
-  default as Toggle,
-  ToggleHandlerProps,
-  Props as ToggleProps,
-} from './Toggle'
+export { default } from './Toggle'