+++ packages/asc-ui/src/index.ts
@@ -104,6 +104,7 @@ import {
 import Tag, { TagStyle } from './components/Tag'
 import TextArea, { TextAreaStyles } from './components/TextArea'
 import TextField, { TextFieldStyles } from './components/TextField'
+import Toggle from './components/Toggle'
 import TopBar, { TopBarStyles } from './components/TopBar'
 import TopTaskLink, { TopTaskLinkStyle } from './components/TopTaskLink'
 import Typography from './components/Typography'
@@ -205,6 +206,7 @@ export {
   Theme,
   ThemeProvider,
   CompactThemeProvider,
+  Toggle,
   TopBar,
   TopTaskLink,
   EditorialPost,